provide a switch for a "live" development mode
this changes the mount of the compose file and the mount point for the debugger.

once switch all changes that you do are directly changed in the container aswell, so you don't have to run the scoper again and restart the compose.

but please be aware you still have to test in scoped mode because that's what we release to the customers.